requirements.txt
djangorestframework
django
drf_yasg